+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.3.0 (February 3, 2016)
 
 - Deprecation warning added for Oy.Element `type` prop. Use `tagName` instead.
-- Added warning for Oy.renderTemplate outputs larger than 100KB.
+- Added warning for Oy.renderTemplate outputs larger than 100KB. See #24
 - Opened up Oy-supported attributes to all Oy.Element components.
 
 ## 0.2.3 (February 1, 2016)