+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 - Deprecation warning added for Oy.Element `type` prop. Use `tagName` instead.
 - Added warning for Oy.renderTemplate outputs larger than 100KB. See #24
 - Opened up Oy-supported attributes to all Oy.Element components.
+- New rules:
+  - EmptyTDRule
+  - FontShorthandRule
 
 ## 0.2.3 (February 1, 2016)